# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: "club-house-project-api"
# during dev, should point to your local machine
host: localhost:8090
basePath: /clubhouse
# 
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
  /drink:
    get:
      description: Returns all the drinks existing in the clubhouse
      # used as the method name of the controller
      operationId: getAll
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              $ref: "#/definitions/DrinkResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /player:
    get:
      description: Get all players
      operationId: getAll
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              $ref: "#/definitions/PlayerResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /player/_create:
    post:
      description: Register a player in database
      operationId: create
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/PlayerResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /player/{player}:
    get:
      description: Get a player by id
      operationId: getById
      parameters:
        - in: path
          name: player
          required: true
          type: string
          minimum: 1
          description: The player ID.
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/PlayerResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /player/{player}/bill:
    get:
      description: Get player bill
      operationId: getBill
      parameters:
        - in: path
          name: player
          required: true
          type: string
          minimum: 1
          description: The player ID.
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/PlayerResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: Add bill to a player
      operationId: updateBill
      parameters:
        - in: path
          name: player
          required: true
          type: string
          minimum: 1
          description: The player ID.
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/PlayerResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /player/{player}/password:
    put:
      description: Update a player password
      operationId: updatePassword
      parameters:
        - in: path
          name: player
          required: true
          type: string
          minimum: 1
          description: The player ID.
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/PlayerResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /player/{player}/drinks:
    delete:
      description: Reset a player bill
      operationId: deleteBill
      parameters:
        - in: path
          name: player
          required: true
          type: string
          minimum: 1
          description: The player ID.
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/PlayerResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
definitions:
  DrinkResponse:
    required:
      - _id
      - label
      - price
      - season
    properties:
      _id:
        type: string
      label:
        type: string
      price:
        type: number
      season:
        type: number
  PlayerResponse:
    required:
      - _id
      - firstname
      - licence
      - admin
      - bill
      - password
    properties:
      _id:
        type: string
      firstname:
        type: string
      licence:
        type: number
      admin:
        type: boolean
      bill:
        type: object
      password:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string